loading todays game from external server, displaying animation while loading and error on failure
